+++ tests/testthat/test-object-s3.R
@@ -50,13 +50,3 @@ test_that("user defined functions override primitives", {
   expect_false(is_s3_generic("c"))
   expect_false(is_s3_method("c"))
 })
-
-test_that("@method overrides auto-detection", {
-  out <- parse_text("
-    #' @export
-    #' @method all.equal data.frame
-    all.equal.data.frame <- function(...) 1
-  ")[[1]]
-
-  expect_equal(s3_method_info(out$object$value), c("all.equal", "data.frame"))
-})